backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
Pillow>=9.5.0
python-decouple>=3.8
dj-database-url>=2.0.0
django-storages>=1.13.2
celery>=5.3.0
channels>=4.0.0
psycopg2-binary>=2.9.6
django-filter>=23.2
python-magic>=0.4.27
